requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-js-asset==2.0.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
sqlparse==0.4.3
